# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface PasswordChangeInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  newPassword: string;
}

export interface GiftRedemptionInput {
  /**
     * @minLength 3
     * @maxLength 40
     */
  code: string;
}

export interface GiftRedemptionResponse {
  ok: boolean;
  amount: number;
  wallet: number;
}

export interface GiftCodeInput {
  /**
     * @minLength 3
     * @maxLength 40
     * @pattern ^[A-Za-z0-9-]+$
     */
  code: string;
  /**
     * @minimum 1
     * @maximum 100000000
     */
  amount: number;
  /**
     * @minimum 1
     * @maximum 100000
     */
  maxRedemptions: number;
  /** @nullable */
  expiresAt?: string | null;
}

export interface GiftCodeUpdate {
  enabled: boolean;
}

export interface GiftCode {
  id: string;
  code: string;
  amount: number;
  maxRedemptions: number;
  redemptionCount: number;
  enabled: boolean;
  createdAt: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface HealthStatus {
  status: string;
}

export interface Settings {
  brand: string;
  currency: string;
  supportHandle?: string;
  telegramUrl?: string;
  airtelNumber?: string;
  mtnNumber?: string;
  payeeName?: string;
  termsText?: string;
}

export interface SettingsInput {
  /** @minLength 1 */
  brand?: string;
  currency?: string;
  supportHandle?: string;
  telegramUrl?: string;
  airtelNumber?: string;
  mtnNumber?: string;
  payeeName?: string;
  /** @maxLength 20000 */
  termsText?: string;
}

export interface Product {
  id: string;
  name: string;
  price: number;
  daily: number;
  total: number;
  days: number;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  price: number;
  /** @minimum 0 */
  daily: number;
  /** @minimum 0 */
  total: number;
  /** @minimum 1 */
  days: number;
}

export interface User {
  id: string;
  phone: string;
  referralCode: string;
  wallet: number;
  totalEarned: number;
  createdAt: string;
  /** @nullable */
  lastCheckin?: string | null;
  banned: boolean;
}

export interface RegisterInput {
  /** @minLength 7 */
  phone: string;
  /** @minLength 8 */
  password: string;
  referralCode?: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  ok: boolean;
  user: User;
}

export interface SessionResponse {
  loggedIn: boolean;
  user?: User | null;
}

export interface OkResponse {
  ok: boolean;
}

export interface AdminWalletActionInput {
  /** @minimum 1 */
  amount: number;
  /** @maxLength 160 */
  note?: string;
}

export interface BanUserInput {
  banned: boolean;
}

export interface UserActionResponse {
  ok: boolean;
  user: User;
}

export interface Purchase {
  id: string;
  userId: string;
  productId: string;
  productName: string;
  amount: number;
  status: string;
  purchasedAt: string;
  earningsCredited?: number;
}

export interface Transaction {
  id: string;
  userId: string;
  type: string;
  amount: number;
  createdAt: string;
}

export interface Activity {
  id: string;
  /** @nullable */
  userId?: string | null;
  type: string;
  createdAt: string;
}

export interface Payment {
  id: string;
  userId: string;
  productId: string;
  productName: string;
  amount: number;
  method: string;
  payerPhone: string;
  transactionId: string;
  status: string;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export interface PaymentInput {
  productId: string;
  method: string;
  amount: number;
  payerPhone: string;
  transactionId: string;
}

export interface PaymentResponse {
  ok: boolean;
  paymentId?: string;
  status?: string;
  payment?: Payment;
}

export interface Withdrawal {
  id: string;
  userId: string;
  amount: number;
  fee: number;
  netAmount: number;
  method: string;
  phone: string;
  status: string;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export interface WithdrawalInput {
  /** @minimum 7000 */
  amount: number;
  method: string;
  phone: string;
}

export interface WithdrawalResponse {
  ok: boolean;
  withdrawal?: Withdrawal;
}

export interface CheckinResponse {
  ok: boolean;
  reward: number;
  user: User;
}

export interface Dashboard {
  user: User;
  purchases: Purchase[];
  transactions: Transaction[];
  canWithdraw: boolean;
}

export interface ReferralMember {
  id: string;
  phone: string;
  level: number;
  createdAt: string;
}

export interface Commission {
  id: string;
  level: number;
  amount: number;
  createdAt: string;
}

export interface ReferralOverview {
  code: string;
  link: string;
  directReferrals: number;
  teamSize: number;
  members: ReferralMember[];
  commissions: Commission[];
}

export interface AdminDashboard {
  users: number;
  products: number;
  purchases: number;
  payments: number;
  withdrawals: number;
  transactions: number;
  walletBalances: number;
  totalDeposited: number;
  totalWithdrawn: number;
  totalInvested: number;
}

export interface AdminReferralSummary {
  commissions: number;
  total: number;
}

export type ReviewInputAction = typeof ReviewInputAction[keyof typeof ReviewInputAction];


export const ReviewInputAction = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface ReviewInput {
  action: ReviewInputAction;
}