# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /account/password:
    post:
      operationId: changeAccountPassword
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChangeInput" }
      responses:
        "200":
          description: Password changed; other sessions revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /account/gift-codes/redeem:
    post:
      operationId: redeemGiftCode
      tags: [account]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GiftRedemptionInput" }
      responses:
        "200":
          description: Gift credited to wallet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GiftRedemptionResponse" }
  /admin/gift-codes:
    get:
      operationId: getAdminGiftCodes
      tags: [admin]
      responses:
        "200":
          description: Gift codes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GiftCode" }
    post:
      operationId: createAdminGiftCode
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GiftCodeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GiftCode" }
  /admin/gift-codes/{id}:
    patch:
      operationId: updateAdminGiftCode
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GiftCodeUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GiftCode" }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 128 }
        newPassword: { type: string, minLength: 8, maxLength: 128 }
    GiftRedemptionInput:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 3, maxLength: 40 }
    GiftRedemptionResponse:
      type: object
      required: [ok, amount, wallet]
      properties:
        ok: { type: boolean }
        amount: { type: integer }
        wallet: { type: number }
    GiftCodeInput:
      type: object
      required: [code, amount, maxRedemptions]
      properties:
        code: { type: string, minLength: 3, maxLength: 40, pattern: "^[A-Za-z0-9-]+$" }
        amount: { type: integer, minimum: 1, maximum: 100000000 }
        maxRedemptions: { type: integer, minimum: 1, maximum: 100000 }
        expiresAt: { type: ["string", "null"], format: date-time }
    GiftCodeUpdate:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
    GiftCode:
      type: object
      required: [id, code, amount, maxRedemptions, redemptionCount, enabled, createdAt, expiresAt]
      properties:
        id: { type: string }
        code: { type: string }
        amount: { type: integer }
        maxRedemptions: { type: integer }
        redemptionCount: { type: integer }
        enabled: { type: boolean }
        createdAt: { type: string }
        expiresAt: { type: ["string", "null"] }
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...