# lib/api-zod/src/generated/types/giftCode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCode {
  id: string;
  code: string;
  amount: number;
  maxRedemptions: number;
  redemptionCount: number;
  enabled: boolean;
  createdAt: string;
  /** @nullable */
  expiresAt: string | null;
}