# lib/api-zod/src/generated/types/giftCodeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCodeInput {
  /**
     * @minLength 3
     * @maxLength 40
     * @pattern ^[A-Za-z0-9-]+$
     */
  code: string;
  /**
     * @minimum 1
     * @maximum 100000000
     */
  amount: number;
  /**
     * @minimum 1
     * @maximum 100000
     */
  maxRedemptions: number;
  /** @nullable */
  expiresAt?: Date | null;
}