# lib/api-zod/src/generated/types/giftRedemptionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftRedemptionInput {
  /**
     * @minLength 3
     * @maxLength 40
     */
  code: string;
}