# lib/api-zod/src/generated/types/giftRedemptionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftRedemptionResponse {
  ok: boolean;
  amount: number;
  wallet: number;
}