# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordChangeInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  newPassword: string;
}